Separate webpack files for building and serving
